// src/api/generated/schemas/deploymentGroup.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * NervesHub API
 * NervesHub firmware update and device management API
 * OpenAPI spec version: 2.0.0
 */
import type { DeploymentGroupConditions } from './deploymentGroupConditions';
import type { DeploymentGroupState } from './deploymentGroupState';
import type { Firmware } from './firmware';

export interface DeploymentGroup {
  id?: number;
  name?: string;
  is_active?: boolean;
  state?: DeploymentGroupState;
  firmware?: Firmware;
  conditions?: DeploymentGroupConditions;
  device_count?: number;
  inserted_at?: string;
  updated_at?: string;
}